pyproject.toml
[project]
name = "ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.11.0",
    "transformers>=5.5.3",
    "requests",
    "Pillow",
]
